+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2.2 (Next)
 ==========
 
+* [#12](https://github.com/dblock/dotnetinstaller/issues/12): Processor architecture filter matching for negated values (using `!`) is incorrect.
 * Displaying processor architecture next to the operating system version in the log file.
 
 2.1 (5/28/2012)